Refactor: render 관련함수 분리(render.js, renderWinner.js)
